Solved issue with incompatible ID type in mySQL during execution spring.jpa.hibernate.ddl-auto=update - expected int by mySQL - expected Long in good practice - casting in CarServiceImpl; improved html

+++ src/main/java/pl/kamil/rentcars/repository/CarRepository.java
@@ -5,7 +5,7 @@ import pl.kamil.rentcars.entity.Car;
 
 import java.util.List;
 
-public interface CarRepository extends JpaRepository<Car, Long> {
+public interface CarRepository extends JpaRepository<Car, Integer> {
     List<Car> findAllByOrderByBrandAsc();
     List<Car> findAllByOrderByModelAsc();
     List<Car> findAllByOrderByVinAsc();
